On the 10Gb branch: Make membuffer handling of lock contention configurable.
We introduce two additional parameters.  The first explicitly sets the
number of cache segments to create -- more segments equals finer grained
locking.  Defaults will still be used if count=0 has been specified.

The second parameter controls how we aquire write locks.  Concurrent reads
are already supported but reads and writes are mutually exclusive for a
given cache segment.  As opposed to writes, reads may invoke time-consuming
callbacks.  To not block the writing thread for a long time, add a flag
that indicated whether to block or not.  In non-blocking mode, if we didn't
get the lock immediately, the object will simply not be put into the cache.

* subversion/include/private/svn_cache.h
  (svn_cache__membuffer_cache_create): add new parameters

* subversion/libsvn_subr/cache-membuffer.c
  (MIN_SEGMENT_SIZE): rename to ...
  (DEFAULT_MIN_SEGMENT_SIZE): ... this
  (MIN_SEGMENT_SIZE,
   MAX_SEGMENT_COUNT): new constants
  (svn_membuffer_t): add allow_blocking_writes flag
  (write_lock_cache): make non-blocking if indicated by flag
  (WITH_WRITE_LOCK): execute EXPR only if we got the write lock
  (svn_cache__membuffer_cache_create): process new parameters

* subversion/libsvn_subr/cache_config.c
  (svn_cache__get_global_membuffer_cache): update
* subversion/tests/libsvn_subr/cache-test.c
  (test_membuffer_cache_basic): update